feat(nav): implemeny all nav for 1 components
